requirements.txt
django>=5.1,<5.2
django-environ>=0.11,<0.12
djangorestframework>=3.15,<3.16
celery[redis]>=5.4,<5.5
redis>=5.2,<5.3
psycopg2-binary>=2.9,<2.10
slack-bolt>=1.21,<1.22
slack-sdk>=3.33,<3.34
gunicorn>=23.0,<24.0
requests>=2.32,<2.33
httpx>=0.28,<0.29
